Move `struct Frame` into own private module
